Features for adding CRUD for a Bulk Discount

+++ db/migrate/20210610212315_create_bulk_discounts.rb
@@ -1,7 +1,7 @@
 class CreateBulkDiscounts < ActiveRecord::Migration[5.2]
   def change
     create_table :bulk_discounts do |t|
-      t.float :percentage_discount
+      t.integer :percentage_discount
       t.integer :quantity_threshold
       t.references :merchant, foreign_key: true
